Fix build: add juce_generate_juce_header, fix sign-conversion warnings
juce_generate_juce_header(SaarthalDelay) generates JuceHeader.h at
configure time so our sources can include it. DelayLine now uses
size_t for buffer indices to eliminate sign-conversion warnings.

+++ src/dsp/DelayLine.h
@@ -1,6 +1,7 @@
 #pragma once
 #include <vector>
 #include <algorithm>
+#include <cstddef>
 
 // Interpolated circular buffer used by all delay engines
 class DelayLine
@@ -8,7 +9,7 @@ class DelayLine
 public:
     void prepare(double sampleRate, double maxDelaySeconds)
     {
-        size = static_cast<int>(sampleRate * maxDelaySeconds) + 2;
+        size = static_cast<size_t>(sampleRate * maxDelaySeconds) + 2;
         buffer.assign(size, 0.0f);
         writePos = 0;
     }
@@ -25,15 +26,15 @@ public:
     float read(float delaySamples) const
     {
         float readPos = static_cast<float>(writePos) - delaySamples;
-        while (readPos < 0) readPos += static_cast<float>(size);
-        int i0 = static_cast<int>(readPos) % size;
-        int i1 = (i0 + 1) % size;
-        float frac = readPos - static_cast<float>(static_cast<int>(readPos));
+        while (readPos < 0.0f) readPos += static_cast<float>(size);
+        size_t i0 = static_cast<size_t>(readPos) % size;
+        size_t i1 = (i0 + 1) % size;
+        float frac = readPos - static_cast<float>(static_cast<size_t>(readPos));
         return buffer[i0] + frac * (buffer[i1] - buffer[i0]);
     }
 
 private:
     std::vector<float> buffer;
-    int size = 0;
-    int writePos = 0;
+    size_t size = 0;
+    size_t writePos = 0;
 };
